pyproject.toml
[project]
name = "openclaw-deepagents-mvp"
version = "0.1.0"
description = "OpenClaw-like personal AI assistant MVP built with LangChain Deep Agents"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "deepagents>=0.5.2",
  "langchain>=1.0.0",
  "langgraph>=1.0.0",
  "langchain-openai>=1.0.0",
  "langchain-anthropic>=1.0.0",
  "langchain-google-genai>=3.0.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "tavily-python>=0.5.0",
]

[project.scripts]
openclaw-da = "openclaw_da.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
